Переименование страниц, настройка кастомных маршрутов и распределение моделей страниц по отдельным файлам
