+++ source/Models/ContestEventsPage/GradeCriterion.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.ContestEventsPage;
+
+public record GradeCriterion(string Identifier, string Name);
