+++ source/Models/ContestEventsPage/ViewModel.cs
@@ -1,0 +1,7 @@
+namespace ConventionGradingSystem.Models.ContestEventsPage;
+
+public record ViewModel(
+    string ContestName,
+    IReadOnlyList<GradeCriterion> ExpertCriterions,
+    IReadOnlyList<GradeCriterion> ParticipantCriterions,
+    IReadOnlyList<ContestEvent> Events);
