+++ source/Models/EventAttendanceForm/Participant.cs
@@ -1,0 +1,8 @@
+namespace ConventionGradingSystem.Models.EventAttendanceForm;
+
+public record Participant(
+    string Identifier,
+    string Name,
+    string Brigade,
+    bool AttendanceMark,
+    bool SpecialMark);
