+++ source/Models/EventAttendanceForm/ViewModel.cs
@@ -4,10 +4,3 @@ public record ViewModel(
     string ContestName,
     string EventName,
     IReadOnlyCollection<Participant> Participants);
-
-public record Participant(
-    string Identifier,
-    string Name,
-    string Brigade,
-    bool AttendanceMark,
-    bool SpecialMark);
