+++ source/Models/ExpertFeedbackForm/FormModel.cs
@@ -3,5 +3,3 @@ namespace ConventionGradingSystem.Models.ExpertFeedbackForm;
 public record FormModel(
     IReadOnlyCollection<Grade> Grades,
     string Note);
-
-public record Grade(string CriterionId, int GradeValue);
