+++ source/Models/ExpertFeedbackForm/Grade.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.ExpertFeedbackForm;
+
+public record Grade(string CriterionId, int GradeValue);
