+++ source/Models/ExpertFeedbackForm/GradeCriterion.cs
@@ -1,0 +1,8 @@
+namespace ConventionGradingSystem.Models.ExpertFeedbackForm;
+
+public record GradeCriterion(
+    string Identifier,
+    string Name,
+    string Description,
+    int MinimalGrade,
+    int MaximalGrade);
