+++ source/Models/ExpertFeedbacksPage/Feedback.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.ExpertFeedbacksPage;
+
+public record Feedback(IReadOnlyDictionary<string, int> Grades, string? Note);
