+++ source/Models/ExpertFeedbacksPage/GradeCriterion.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.ExpertFeedbacksPage;
+
+public record GradeCriterion(string Identifier, string Name);
