+++ source/Models/ExpertFeedbacksPage/ViewModel.cs
@@ -1,0 +1,7 @@
+namespace ConventionGradingSystem.Models.ExpertFeedbacksPage;
+
+public record ViewModel(
+    string ContestName,
+    string EventName,
+    IReadOnlyList<GradeCriterion> Criterions,
+    IReadOnlyList<Feedback> Feedbacks);
