--- source/Models/Login/FormModel.cs
@@ -1,3 +1,0 @@
-namespace ConventionGradingSystem.Models.Login;
-
-public record FormModel(string Secret);
