+++ source/Models/LoginForm/FormModel.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.LoginForm;
+
+public record FormModel(string Secret);
