+++ source/Models/MainPage/Contest.cs
@@ -1,0 +1,6 @@
+namespace ConventionGradingSystem.Models.MainPage;
+
+public record Contest(
+    string Identifier,
+    string Name,
+    int EventsQuantity);
