+++ source/Models/MainPage/Team.cs
@@ -1,0 +1,8 @@
+namespace ConventionGradingSystem.Models.MainPage;
+
+public record Team(
+    string Name,
+    int MembersQuantity,
+    int MembersRegistrationsQuantity,
+    int MembersMarksQuantity,
+    int SpecialMarksQuantity);
