+++ source/Models/MainPage/ViewModel.cs
@@ -1,0 +1,5 @@
+namespace ConventionGradingSystem.Models.MainPage;
+
+public record ViewModel(
+    IReadOnlyList<Contest> Contests,
+    IReadOnlyList<Team> Teams);
