+++ source/Models/ParticipantFeedbackForm/FormModel.cs
@@ -3,5 +3,3 @@ namespace ConventionGradingSystem.Models.ParticipantFeedbackForm;
 public record FormModel(
     IReadOnlyCollection<Grade> Grades,
     string Note);
-
-public record Grade(string CriterionId, int GradeValue);
