+++ source/Models/ParticipantFeedbackForm/Grade.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.ParticipantFeedbackForm;
+
+public record Grade(string CriterionId, int GradeValue);
