+++ source/Models/ParticipantFeedbackForm/GradeCriterion.cs
@@ -1,0 +1,8 @@
+namespace ConventionGradingSystem.Models.ParticipantFeedbackForm;
+
+public record GradeCriterion(
+    string Identifier,
+    string Name,
+    string Description,
+    int MinimalGrade,
+    int MaximalGrade);
