+++ source/Models/ParticipantFeedbacksPage/Feedback.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.ParticipantFeedbacksPage;
+
+public record Feedback(IReadOnlyDictionary<string, int> Grades, string? Note);
