+++ source/Models/ParticipantFeedbacksPage/GradeCriterion.cs
@@ -1,0 +1,3 @@
+namespace ConventionGradingSystem.Models.ParticipantFeedbacksPage;
+
+public record GradeCriterion(string Identifier, string Name);
