+++ source/Models/ParticipantFeedbacksPage/ViewModel.cs
@@ -1,0 +1,7 @@
+namespace ConventionGradingSystem.Models.ParticipantFeedbacksPage;
+
+public record ViewModel(
+    string ContestName,
+    string EventName,
+    IReadOnlyList<GradeCriterion> Criterions,
+    IReadOnlyList<Feedback> Feedbacks);
