+++ source/Pages/AccessDeniedPage.cshtml
@@ -1,4 +1,4 @@
-﻿@page
+@page "denied"
 
 @{
     ViewData["Title"] = "Доступ запрещён";
