+++ source/Pages/ExpertFeedbackForm.cshtml
@@ -1,4 +1,4 @@
-@page "{eventId}"
+@page "/expert-feedback/{eventId}"
 @model ConventionGradingSystem.Pages.ExpertFeedbackFormModel
 @using ConventionGradingSystem.Models.ExpertFeedbackForm
