+++ source/Pages/ExpertFeedbacksPage.cshtml
@@ -1,5 +1,5 @@
-@page "{eventId}"
-@model ConventionGradingSystem.Pages.ExpertFeedbacksModel
+@page "/expert-feedbacks/{eventId}"
+@model ConventionGradingSystem.Pages.ExpertFeedbacksPageModel
 
 @{
     ViewData["Title"] = "Комментарии экспертов";
