+++ source/Pages/LoginForm.cshtml
@@ -1,5 +1,5 @@
-@page
-@model ConventionGradingSystem.Pages.LoginModel
+@page "/login"
+@model ConventionGradingSystem.Pages.LoginFormModel
 
 @{
     ViewData["Title"] = "Вход в систему";
