+++ source/Pages/ParticipantFeedbackForm.cshtml
@@ -1,4 +1,4 @@
-@page "{eventId}"
+@page "/participant-feedback/{eventId}"
 @model ConventionGradingSystem.Pages.ParticipantFeedbackFormModel
 @using ConventionGradingSystem.Models.ParticipantFeedbackForm
