+++ source/Pages/ParticipantFeedbacksPage.cshtml
@@ -1,5 +1,5 @@
-@page "{eventId}"
-@model ConventionGradingSystem.Pages.ParticipantFeedbacksModel
+@page "/participant-feedbacks/{eventId}"
+@model ConventionGradingSystem.Pages.ParticipantFeedbacksPageModel
 
 @{
     ViewData["Title"] = "Комментарии участников";
