+++ source/Pages/Shared/_Layout.cshtml
@@ -270,7 +270,7 @@
         <div class="title">
             @if (@ViewBag.AllowNavigation == true)
             {
-                <a class="title" href="/Index">Система оценивания Слёта СПбСО</a>
+                <a class="title" href="/">Система оценивания Слёта СПбСО</a>
             }
             else
             {
